% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java.R
\name{add_java_env}
\alias{add_java_env}
\title{Set some environment variables required by some packages.}
\usage{
add_java_env(home = "C:\\\\Program Files\\\\Java\\\\jdk1.8.0_66\\\\bin",
  javac = home,
  jvm = "C:\\\\Program Files\\\\Java\\\\jdk1.8.0_66\\\\jre\\\\bin\\\\server",
  overwrite = FALSE)
}
\arguments{
\item{home}{Path for JAVA_HOME}

\item{javac}{Path to javac (probably same as HOME)}

\item{jvm}{Path to jvm.dll (/path/to/jdk/jre/bin/server/)}

\item{overwrite}{If true, overwrite JAVA_HOME}
}
\description{
Set some environment variables required by some packages.
}

