% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{binRead}
\alias{binRead}
\title{binRead}
\usage{
binRead(fName)
}
\arguments{
\item{fName}{Filename}
}
\value{
bytes
}
\description{
Can use this to compare compressed archives for example
}
\author{
bluefish (SO tag)
}
\seealso{
\link{http://stackoverflow.com/questions/34189716/test-if-compressed-archives-contain-same-data/34191638#34191638}
}

