% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{blankDF}
\alias{blankDF}
\title{Create empty data.frame with all possible names of supplied arguments}
\usage{
blankDF(...)
}
\arguments{
\item{...}{data.frame types we want the names of}
}
\value{
empty data.frame with union of names of passed objects
}
\description{
Create empty data.frame with all possible names of supplied arguments
}

