% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{captureInputs}
\alias{captureInputs}
\title{Store input values, so when partials are rendered values can be restored}
\usage{
captureInputs(prefix, rvals, input)
}
\arguments{
\item{prefix}{UI prefixes}

\item{rvals}{global reactive values}

\item{input}{global input}
}
\description{
Store input values, so when partials are rendered values can be restored
}

