% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{class-length-checks}
\alias{class-length-checks}
\alias{check_logical_length_one}
\alias{check_character_length_one}
\alias{check_null_or_character_length_one}
\title{Function class and length checks}
\usage{
check_logical_length_one(x)

check_character_length_one(x)

check_null_or_character_length_one(x)
}
\arguments{
\item{x}{The value to be checked}
}
\value{
TRUE, invisibly, if the check succeeds
}
\description{
Often in R package development, the need arises to verify values are of both
a specific class and length, with null's allowed conditionally. These are a
collection of such functions.
}
\examples{

# logical: length-one
check_logical_length_one(TRUE)

try(check_logical_length_one(rep(FALSE, 2)))

try(check_logical_length_one("TRUE"))


# character: length-one
check_character_length_one("x")

try(check_character_length_one(NULL))


# character: NULL or length-one
check_null_or_character_length_one("x")
check_null_or_character_length_one(NULL)

try(check_null_or_character_length_one(TRUE))
}
