% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{clean_src}
\alias{clean_src}
\title{Clean all tar files from source dir}
\usage{
clean_src(pattern = "\\\\.gz$")
}
\arguments{
\item{pattern}{file extension pattern of files to remove}
}
\description{
Clean all tar files from source dir
}

