% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_github_pkgdown}
\alias{create_github_pkgdown}
\title{Create a Github Repo}
\usage{
create_github_pkgdown(whoami, open = FALSE)
}
\arguments{
\item{whoami}{The result of \link[gh:gh_whoami]{gh::gh_whoami}}

\item{open}{Should a web browser be opened to the new pkgdown site? Default is \code{FALSE}.}
}
\value{
The URL of the newly created pkgdown site
}
\description{
From inside a project directory, for a project that has a repository on Github, call this
function to create a pkgdown site hosted on Github Pages, and automatically deployed using
Github Actions.
}
\details{
This function uses the user-provided response from \code{\link[gh:gh_whoami]{gh::gh_whoami()}} to identify the Github
repo. It will also update the \code{URL} in the \code{DESCRIPTION} to include the pkgdown site URL.
}
