% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_github_repo}
\alias{create_github_repo}
\title{Create a Github Repo}
\usage{
create_github_repo(
  whoami,
  github_private = FALSE,
  add_cran_check_badge = TRUE,
  open = FALSE
)
}
\arguments{
\item{whoami}{The result of \link[gh:gh_whoami]{gh::gh_whoami}}

\item{github_private}{Should the github repository be private? Default \code{FALSE}.}

\item{add_cran_check_badge}{Should the R-CMD-CHECK be added to the README? Default is \code{TRUE}.}

\item{open}{Should a web browser be opened to the new Github repository? Default is \code{FALSE}.}
}
\value{
The URL of the newly created Github repository
}
\description{
From inside a project directory, call this function to create a repo in your a Github
repository.
}
\details{
This function uses the user-provided response from \code{\link[gh:gh_whoami]{gh::gh_whoami()}} to create a Github repo. It
will also paste the description from the \code{DESCRIPTION} file into the repository description on
Github. It will also update the \code{URL} and \code{BugReports} fields in \code{DESCRIPTION} with the URL's
from the newly created repo. If \code{add_cran_check_badge} is set to \code{TRUE}, the R-CMD-CHECK badge
will be added to the README.
}
