% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{Create a New R Project}
\usage{
create_project(
  dir,
  package_title = NULL,
  package_description = NULL,
  package_license = NULL,
  package_version = NULL,
  github_use = TRUE,
  github_pkgdown = TRUE,
  github_private = FALSE,
  open = TRUE
)
}
\arguments{
\item{dir}{A character string specifying the name of the directory to create the project in.
The directory will also be used as the package name.}

\item{package_title}{Optional title for the package, automatically capitalized if provided.}

\item{package_description}{Optional description for the package.}

\item{package_license}{Optional license for the package, defaults to MIT when not specified. If
specified, a proprietary license using the provided \code{package_license} is used.}

\item{package_version}{Optional version number for the package, e.g. \code{"0.1.0"}.}

\item{github_use}{Logical indicating whether to create a new GitHub repository and link it with
the project. Defaults to TRUE.}

\item{github_pkgdown}{Logical indicating whether to set up GitHub Pages for pkgdown
documentation. Requires \code{github_use} to be TRUE.}

\item{github_private}{Logical indicating whether the created repository should be private on
GitHub. Only applicable if \code{github_use} is TRUE.}

\item{open}{Logical indicating whether to open the newly created project with RStudio
(requires \code{rstudioapi}). Defaults to TRUE.}
}
\value{
The path to the newly created project directory as an fs_path object.
}
\description{
Creates a new R project with basic structure and configurations for development, version control,
and optional GitHub/pkgdown integration.
}
