% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{fill_spaces}
\alias{fill_spaces}
\title{Fill strings so all are the same length as width
by filling with additional spaces.}
\usage{
fill_spaces(lines, width, fill = c("random", "right", "left"))
}
\arguments{
\item{lines}{Vector of strings}

\item{width}{Width to make strings}

\item{fill}{Preference of where to add additional spaces (default is random)}
}
\value{
vector of justified strings
}
\description{
Fill strings so all are the same length as width
by filling with additional spaces.
}

