% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{findData}
\alias{findData}
\title{Find paths to data in specified diretories}
\usage{
findData(locs, files, throw_error = TRUE)
}
\arguments{
\item{locs}{directories to search}

\item{files}{names of files to look for}

\item{throw_error}{if specified, throw errow when not found}
}
\value{
list of two character vectors: found files, missed files
}
\description{
Find paths to data in specified diretories
}

