% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{findPacks}
\alias{findPacks}
\title{Gathers requires/libraries}
\usage{
findPacks(dirs = ".", fnames = NULL, pattern = ".*\\\\.[Rr]+",
  recursive = TRUE)
}
\arguments{
\item{dirs}{Directories to search}

\item{fnames}{Filenames to search in (optional)}

\item{pattern}{File extension of files to search}

\item{recursive}{If true, search recursively}
}
\description{
Gathers requires/libraries
}

