% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{findParts}
\alias{findParts}
\title{Return a list of partials/controllers from subdirectories}
\usage{
findParts(ids = c("partials", "controllers"))
}
\arguments{
\item{ids}{Names of directories to search in}
}
\value{
list of partials and controllers
}
\description{
Return a list of partials/controllers from subdirectories
}

