% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{grepInOrder}
\alias{grepInOrder}
\title{Grep indices ordered by year}
\usage{
grepInOrder(coln, yrs, dat)
}
\arguments{
\item{coln}{Columns}

\item{yrs}{Years}

\item{dat}{data}
}
\description{
Grep indices ordered by year
}

