% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{install_packages}
\alias{install_packages}
\title{Install, keep source, untar, and unlink, and make a tags file. 
Assumes an environment variable 'R_SRC' has been set to the directory
where source should be stored.}
\usage{
install_packages(pkg, tag = TRUE, ...)
}
\arguments{
\item{pkg}{Vector of package names (strings)}

\item{tag}{Logical to use rtags to tag source (default TRUE).}

\item{...}{Passed to install.packages}
}
\description{
Install, keep source, untar, and unlink, and make a tags file. 
Assumes an environment variable 'R_SRC' has been set to the directory
where source should be stored.
}

