% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{is_writeable}
\alias{is_writeable}
\title{Check file write permission.}
\usage{
is_writeable(x)
}
\arguments{
\item{x}{File path}
}
\description{
Wrapper for file.access to test file write permissions.
From \code{dplyr}
}

