% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{justify}
\alias{justify}
\title{Justify vector of strings to all be the same length as width
by filling with additional spaces.}
\usage{
justify(lines, width, fill = c("random", "right", "left"))
}
\arguments{
\item{lines}{Vector of strings}

\item{width}{Width to make strings}

\item{fill}{Preference of where to add additional spaces (default is random)}
}
\value{
Justified strings
}
\description{
Justify vector of strings to all be the same length as width
by filling with additional spaces.
}

