% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{levOrder}
\alias{levOrder}
\title{If the underlying variable is numeric, like PPLOT,
lets order the factor numerically, otherwise use a standard}
\usage{
levOrder(v)
}
\arguments{
\item{v}{Variable}
}
\value{
Order of levels
}
\description{
If the underlying variable is numeric, like PPLOT,
lets order the factor numerically, otherwise use a standard
}

