% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afs.R
\name{load_docs}
\alias{load_docs}
\title{Copy documents over from AFS}
\usage{
load_docs(docs, outdir)
}
\arguments{
\item{docs}{Documents to copy}

\item{outdir}{Directory to copy to (or an "./inst/docs/" directory is created)}
}
\description{
Copy documents over from AFS
}

