% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{make_factor}
\alias{make_factor}
\title{Make factors from data column}
\usage{
make_factor(col, data, drop = TRUE)
}
\arguments{
\item{col}{Column}

\item{data}{Data}

\item{drop}{Drop NAs}
}
\description{
I don't think I ever use this?  Since data.tables store characters 
just as efficiently with the string cache...
}

