% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{makePlotContainers}
\alias{makePlotContainers}
\title{Make dynamic plot containers}
\usage{
makePlotContainers(n, ncol = 2, prefix = "plot", height = 100,
  width = "100\%", ...)
}
\arguments{
\item{n}{Number of containers}

\item{ncol}{Number of columns}

\item{prefix}{prefix for container name}

\item{height}{height (css)}

\item{width}{width (css)}

\item{...}{unused}
}
\description{
Make dynamic plot containers
}

