% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{pol2cart}
\alias{pol2cart}
\title{Convert polar to cartesian coordinates}
\usage{
pol2cart(r, theta, deg = FALSE, recycle = FALSE)
}
\arguments{
\item{r}{Distance}

\item{theta}{Angle (default radians)}

\item{deg}{If specified, then theta is in degrees}

\item{recycle}{If specified, recycles \code{r}}
}
\value{
Matrix of x,y coordinates
}
\description{
Convert polar to cartesian coordinates
}

