% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{prettify}
\alias{prettify}
\title{Make nice for markdown}
\usage{
prettify(x, type = c("each", "all"))
}
\arguments{
\item{x}{input vector}

\item{type}{'each' wraps each element in ``, 'all' wraps the whole vector in ``}
}
\description{
Make nice for markdown
}

