% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{pretty_list}
\alias{pretty_list}
\title{Convert input to list}
\usage{
pretty_list(x, use.na = FALSE, type = c("ul", "ol"))
}
\arguments{
\item{x}{Input vector}

\item{use.na}{If true, leave NA in result}

\item{type}{List type (ordered/unordered)}
}
\description{
Print vector as a (ordered/unordered) list
}

