% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderPlots}
\alias{renderPlots}
\title{Render dynamic plot outputs}
\usage{
renderPlots(n, input, output, prefix = "plot")
}
\arguments{
\item{n}{Number of outpus}

\item{input}{global input}

\item{output}{global output}

\item{prefix}{prefix of plot ids}
}
\description{
Render dynamic plot outputs
}

