% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\name{rep_vars}
\alias{rep_vars}
\title{Function to replace variables in function body
expr is `body(f)`, keyvals is a lookup table for replacements}
\usage{
rep_vars(expr, keyvals)
}
\arguments{
\item{expr}{body of function for example, ie `body(func)`}

\item{keyvals}{named lookup}
}
\value{
new expression with substituted variables
}
\description{
Function to replace variables in function body
expr is `body(f)`, keyvals is a lookup table for replacements
}
\examples{
# replace the formals of a function and its body variables
# http://stackoverflow.com/questions/33850219/change-argument-names-inside-a-function-r
\dontrun{
  f <- function(x, y) -x^2 + x + -y^2 + y
  newvals <- c('x'='x0', 'y'='y0')      # named lookup vector
  newbod <- rep_vars(body(f), newvals)  # create new body
  # Rename the formals, and update the body
  formals(f) <- pairlist(x0=bquote(), y0=bquote())
  body(f) <- newbod
}
}

