% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afs.R
\name{setup_data}
\alias{setup_data}
\title{Copy data from AFS and create roxygen documentation.}
\usage{
setup_data(files, basedir = ".", descriptions, capitalize = TRUE,
  datadir = "data", docdir = "R", add_tables = TRUE, ...)
}
\arguments{
\item{files}{Data files to get from AFS}

\item{basedir}{base directory (default to '.')}

\item{descriptions}{Optional data descriptions}

\item{capitalize}{If true, capitalize the variable names}

\item{datadir}{Name of output directory for data (Default to 'data')}

\item{docdir}{Name of output directory for files produced by datadoc::describe (default 'R')}

\item{add_tables}{If true, add tables to documentation.}

\item{...}{passed to datadoc::describe}
}
\description{
Copy data from AFS and create roxygen documentation.
}

