% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{Sys.which2}
\alias{Sys.which2}
\title{Modification of Sys.which that can find the ImageMagik convert.exe}
\usage{
Sys.which2(cmd)
}
\arguments{
\item{cmd}{System command to find}
}
\description{
Modification of Sys.which that can find the ImageMagik convert.exe
}
\seealso{
\link{http://stackoverflow.com/questions/34030087/how-to-stop-sys-which-from-sucking-on-windows/34031214?noredirect=1#34031214}
}

