% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{justify}
\alias{justify}
\title{Justify strings like in open office or excel so each line is
the exact same length (this doesn't do it).}
\usage{
justify(string, width = getOption("width"), fill = c("random", "right",
  "left"))
}
\arguments{
\item{string}{input string}

\item{width}{width of justified outcome.}

\item{fill}{where additional spaces should be added (default random)}
}
\value{
justifed string
}
\description{
Justify strings like in open office or excel so each line is
the exact same length (this doesn't do it).
}

