% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{levOrder}
\alias{levOrder}
\title{Level ordering}
\usage{
levOrder(v, include.empty = TRUE, empty = c("back", "front"))
}
\arguments{
\item{v}{Variable to convert to factor}

\item{include.empty}{If specified, include empty strings as factors}

\item{empty}{If \code{include.empty} is true, specifies location of empty level.}
}
\value{
Vector containing order of levels.
}
\description{
Create factor levels in a standard way
}
\details{
If the underlying variable is numeric, like PPLOT,
orders the factor numerically, otherwise uses a standard sort
}

