% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{Sys.which2}
\alias{Sys.which2}
\title{Modification of Sys.which that can find the ImageMagik convert.exe}
\usage{
Sys.which2(cmd)
}
\arguments{
\item{cmd}{System command to find}
}
\description{
Find the correct .exe on Windows.
}
\seealso{
\link[stackoverflow]{http://stackoverflow.com/questions/34030087/how-to-find-correct-executable-with-sys-which-on-windows}
}

